+++ ui/src/components/Broker/Broker.tsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react'
-import { Switch, Route, useRouteMatch } from 'react-router-dom'
+import { Switch, Route, useRouteMatch, NavLink } from 'react-router-dom'
 
 import { useLedger, useParty } from '@daml/react'
+import { Menu } from 'semantic-ui-react'
+
 
 import { AssetDeposit } from '@daml.js/da-marketplace/lib/DA/Finance/Asset'
 import { RegisteredBroker } from '@daml.js/da-marketplace/lib/Marketplace/Registry'
-import { RegisteredInvestor } from '@daml.js/da-marketplace/lib/Marketplace/Registry'
 
 import { BrokerInvitation } from '@daml.js/da-marketplace/lib/Marketplace/Broker'
 import { CustodianRelationship } from '@daml.js/da-marketplace/lib/Marketplace/Custodian'
@@ -29,7 +30,8 @@ import LandingPage from '../common/LandingPage'
 import Wallet from '../common/Wallet'
 
 import BrokerOrders from './BrokerOrders'
-import BrokerCustomers from './BrokerCustomers'
+import BrokerCustomerHoldings from './BrokerCustomerHoldings'
+import AddBrokerCustomer from './AddBrokerCustomer'
 
 type Props = {
     onLogout: () => void;
@@ -43,6 +45,7 @@ const Broker: React.FC<Props> = ({ onLogout }) => {
 
     const registeredBroker = useContractQuery(RegisteredBroker);
     const allCustodianRelationships = useContractQuery(CustodianRelationship);
+    const brokerCustomers = useContractQuery(BrokerCustomer);
     const allDeposits = useContractQuery(AssetDeposit);
     const notifications = useDismissibleNotifications();
 
@@ -97,9 +100,24 @@ const Broker: React.FC<Props> = ({ onLogout }) => {
                                  name={registeredBroker[0]?.contractData.name || broker}
                                  items={[
                                     {to: `${url}/wallet`, label: 'Wallet', icon: <WalletIcon/>},
-                                    {to: `${url}/orders`, label: 'Orders', icon: <OrdersIcon/>},
-                                    {to: `${url}/customers`, label: 'Customers', icon: <UserIcon/>}
-                                 ]}/>
+                                    {to: `${url}/orders`, label: 'Orders', icon: <OrdersIcon/>}
+                                 ]}>
+                        <Menu.Menu className='sub-menu'>
+                            <Menu.Item>
+                                <p className='p2'>Clients:</p>
+                            </Menu.Item>
+                            {brokerCustomers.map(customer => (
+                                <Menu.Item
+                                    className='sidemenu-item-normal'
+                                    as={NavLink}
+                                    to={`${url}/customer/${customer.contractId}`}
+                                    key={customer.contractId}
+                                >
+                                    <p>{customer.contractData.brokerCustomer}</p>
+                                </Menu.Item>
+                            ))}
+                        </Menu.Menu>
+                    </RoleSideNav>
 
     const brokerScreen =
         <div className='broker'>
@@ -120,8 +138,11 @@ const Broker: React.FC<Props> = ({ onLogout }) => {
                             </FormErrorHandled>
                         }
                         marketRelationships={
-                            <MarketRelationships role={MarketRole.BrokerRole}
-                                                custodianRelationships={allCustodianRelationships}/>}
+                                <>
+                                <MarketRelationships role={MarketRole.BrokerRole}
+                                                    custodianRelationships={allCustodianRelationships}/>
+                                <AddBrokerCustomer/>
+                            </>}
                         sideNav={sideNav}
                         notifications={notifications}
                         onLogout={onLogout}/>
@@ -140,11 +161,11 @@ const Broker: React.FC<Props> = ({ onLogout }) => {
                         deposits={allDeposits}
                         onLogout={onLogout}/>
                 </Route>
-                <Route path={`${path}/customers`}>
-                    <BrokerCustomers
+
+                <Route path={`${path}/customer/:customerId`}>
+                    <BrokerCustomerHoldings
                         sideNav={sideNav}
-                        onLogout={onLogout}
-                        />
+                        onLogout={onLogout}/>
                 </Route>
             </Switch>
         </div>
