+++ ui/src/components/Broker/BrokerCustomerHoldings.scss
@@ -1,0 +1,22 @@
+@mixin broker-customer-holdings {
+    .broker-customer-holdings {
+        .title {
+            display: flex;
+            flex-direction: row;
+            align-items: center;
+
+            h2 {
+                margin-right: $spacing-l;
+            }
+
+            a {
+                display: flex;
+                align-items: center;
+
+                .icon {
+                    margin-right: $spacing-xs;
+                }
+            }
+        }
+    }
+}
