--- ui/src/components/Broker/BrokerCustomers.scss
@@ -1,10 +1,0 @@
-@mixin broker-customers {
-    .broker-customers {
-        height: 100%;
-        .title {
-            display: flex;
-            flex-direction: row;
-            align-items: center;
-        }
-    }
-}
